feat: Add server and client configuration files
Add .gitignore files for the server and client directories to ignore node_modules, .env, and dist/ directories.

Add postcss.config.js file to configure PostCSS plugins for the client.

Add jsconfig.json file to configure the compiler options and paths for the client.

Add NotFound component to handle non-existent pages in the client.

Add cn function to the utils.js file in the client to merge class names using clsx and tailwind-merge.

Add AdminFeatures and UnauthPage components to the client.

Add Feature and paypal modules to the server.

Add AdminOrders, Skeleton, and index.html files to the client.

Add components.json file to configure component aliases in the client.

Add Product, Address, Review, feature-routes, review-routes, and products-routes modules to the server.

Add index.css and layout.jsx files to the shopping-view folder in the client.

Add main.jsx file to render the App component in the client.

Add cart-routes module to the server.